pyproject.toml
[tool.poetry]
name = "dynamic-rag"
version = "0.1.0"
description = "Industry-grade FastAPI RAG system with Ollama, Qdrant, and PostgreSQL"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
ollama = "^0.1.6"
qdrant-client = "^1.7.0"
redis = "^5.0.1"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pytesseract = "^0.3.10"
pdf2image = "^1.17.0"
pypdf2 = "^3.0.1"
pillow = "^10.2.0"
langchain-text-splitters = "^0.0.1"
jinja2 = "^3.1.3"
python-dotenv = "^1.0.0"
celery = {extras = ["redis"], version = "^5.3.4"}
tenacity = "^8.2.3"
httpx = "^0.26.0"
asyncpg = "^0.29.0"
aioredis = "^2.0.1"
rank-bm25 = "^0.2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
black = "^24.1.1"
ruff = "^0.1.14"
mypy = "^1.8.0"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
